Add files via upload
consists of two ECU’s. The first ECU called
HMI responsible for interfacing with the user
and the second ECU called control ECU
which is responsible for the system
operations and control.In the project I
implemented the following drivers;( Keypad,
LCD, DC Motor, UART, Timer, I2C,External
EEPROM).